+++ programs/dynamic-fee-sharing/src/instructions/ix_initialize_fee_vault_pda.rs
@@ -80,7 +80,7 @@ pub fn handle_initialize_fee_vault_pda(
         &ctx.accounts.base.key,
         ctx.bumps.fee_vault,
         FeeVaultType::PdaAccount.into(),
-        params.mutable_flag,
+        params.mutable_flag.into(),
     )?;
 
     emit_cpi!(EvtInitializeFeeVault {